relay-server/requirements.txt
fastapi>=0.110
uvicorn>=0.27
websockets>=12.0
httpx>=0.27
livekit-api>=0.7
